requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
beautifulsoup4==4.9.3
boto==2.49.0
boto3==1.16.5
botocore==1.19.5
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Collectfast==0.2.3
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssutils==1.0.2
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.3.0
django-redis-cache==1.6.5
django-s3-upload==0.3.1
django-sendgrid-v5==0.9.0
django-storages==1.10.1
django-storages-redux==1.3
djangorestframework==3.11.0
docutils==0.15.2
future==0.16.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
jmespath==0.10.0
msgpack==0.6.2
msgpack-python==0.5.6
pathspec==0.5.9
Pillow==7.1.2
promise==2.3
psycopg2==2.7.7
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyHamcrest==2.0.2
pynliner==0.8.0
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-decouple==3.3
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
requests==2.24.0
Rx==1.6.1
s3transfer==0.3.3
semantic-version==2.5.0
sendgrid==6.4.7
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.11.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
termcolor==1.1.0
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.11
wcwidth==0.1.9
whitenoise==5.1.0
zope.interface==5.1.0
